requirements.txt
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-MySQLdb==2.0.0
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
Jinja2==3.1.4
mariadb==1.1.10
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
mysqlclient==2.2.4
packaging==24.1
PyJWT==2.8.0
Werkzeug==3.0.3
